Extracting URLHint to separate module
