+++ dapps/src/apps/mod.rs
@@ -20,6 +20,7 @@ use proxypac::ProxyPac;
 use parity_dapps::WebApp;
 
 mod fs;
+pub mod urlhint;
 pub mod fetcher;
 pub mod manifest;
